new TaskHandlerClass, changed old to ScheduleGenerator
