+++ server/DbHandler.js
@@ -1,4 +1,5 @@
 require("dotenv").config()
+const TaskHandler = require('./TaskHandler');
 const { MongoClient, ServerApiVersion, ObjectId } = require('mongodb');
 //Mongodb doc https://www.mongodb.com/docs/drivers/node/current/
 //Fundamentals section will be your friend
@@ -13,6 +14,7 @@ class DbHandler {
     #db;
     #usersCollection;
     #tasksCollection;
+    #taskHandler;
 
     constructor(){     
         this.#uri = process.env.MONGO_CONNECTION;
@@ -29,6 +31,7 @@ class DbHandler {
         //create connection to each collection for use in routes
         this.#usersCollection = this.#db.collection('Users')
         this.#tasksCollection = this.#db.collection('Tasks') 
+        this.#taskHandler = new TaskHandler();
     }
 
     loginVerify(req){
@@ -87,24 +90,15 @@ class DbHandler {
             try{
                 this.#usersCollection.find({ Username: req.body.Username}).toArray().then(info => {
                     const tempInfo = info;
-                    //console.log('passed find operation', info);
-                    //console.log('active tasks', info[0].ActiveTasks);
-                    //console.log('inactive tasks', info[0].InactiveTasks);
-                    //console.log('recurring', info[0].RecurringTasks);
-                    //console.log('schedules: ', info[0].Schedules);
                     // populate a list with all IDs from info.activeTasks, info.inactiveTasks, info.recurringTasks, info.schedules
                     var tasksList = {};
                     this.#tasksCollection.find({"_id": {$in: tempInfo[0].ActiveTasks} }).toArray().then(info => {
-                        //console.log('tasksCollection return active', info);
                         tasksList["ActiveTasks"] = info;
                         this.#tasksCollection.find({"_id": {$in: tempInfo[0].InactiveTasks} }).toArray().then(info=> {
-                            //console.log('tasksCollection return inactive', info);
                             tasksList["InactiveTasks"] = info;
                             this.#tasksCollection.find({"_id": {$in: tempInfo[0].RecurringTasks} }).toArray().then(info=> {
-                                //console.log('tasksCollection return recurring', info);
                                 tasksList["RecurringTasks"] = info; 
                                 tasksList["Schedules"] = tempInfo[0].Schedules;
-                                //console.log('Final List:', tasksList)
                                 resolve(tasksList);  
                             })           
                         })
@@ -128,7 +122,7 @@ class DbHandler {
         })
     }
 
-    saveTask(req){
+    async saveTask(req){
         return new Promise((resolve, reject) => {
             var task = req.body.Task;
             console.log('savetask ', task);
@@ -137,32 +131,8 @@ class DbHandler {
                 try{
                     tasksCollection.replaceOne({"_id":tempId},task).then(info => {
                         usersCollection.find({"Username":req.body.Username}).toArray().then(user =>{
-                            let recKeyList = Object.keys(task['Recurrence']);
-                            let recFlag = false;
-                            for (var i = 0; i < user[0]['RecurringTasks'].length; i++){
-                                if (String(user[0]['RecurringTasks'][i]) == String(tempId)){
-                                    user[0]['RecurringTasks'].splice(i,1);
-                                    break;
-                                }
-                            }
-                            for(var i = 0; i  < user[0]['InactiveTasks'].length; i++){
-                                if (String(user[0]['InactiveTasks'][i]) == String(tempId)){
-                                    user[0]['InactiveTasks'].splice(i,1);
-                                    break;
-                                }
-                            }
-                            for (var i = 0; i < recKeyList.length; i++){
-                                if (task['Recurrence'][recKeyList[i]]){
-                                    user[0]['RecurringTasks'].push(tempId);
-                                    recFlag = true;
-                                    break;
-                                }
-                            }
-                            if (recFlag == false){
-                                user[0]['InactiveTasks'].push(tempId);
-
-                            }
-                            usersCollection.replaceOne({"_id":user[0]["_id"]},user);
+                            newUser = this.#taskHandler.updateTaskOnUser(user[0],task);
+                            usersCollection.replaceOne({"_id":user[0]["_id"]},newUser);
                             //check for completion of replace???
                         })
                     })
@@ -176,13 +146,9 @@ class DbHandler {
                 delete task["Id"];
                     this.#tasksCollection.insertOne(task).then(info => {
                         task["Id"] = String(info.insertedId);
-                        //console.log('debug info', info)
-                        //console.log('debug', task["_id"]);
                         this.#usersCollection.find({"Username":req.body.Username}).toArray().then(user =>{
-                            //console.log('debug user', user);
-                            user[0]['InactiveTasks'].push(task["_id"]);
-                            //console.log('user debug 2', user);
-                            this.#usersCollection.replaceOne({"_id":user[0]["_id"]},user[0]);
+                            user = this.#taskHandler.addNewTaskToInactive(user[0],task);
+                            this.#usersCollection.replaceOne({"_id":user[0]["_id"]},user);
                         })
                         resolve(task);
                     })
